requirements.txt
Django==4.0.6
psycopg2-binary==2.9.3
django-minio-storage==0.3.10
djangorestframework==3.13.1
